refactor: Change Dir Name (#157)
- res -> response
- req -> request
