--- presentation/src/main/java/way/presentation/feed/vo/req/FeedRequestVo.java
@@ -1,5 +1,0 @@
-package way.presentation.feed.vo.req;
-
-public class FeedRequestVo {
-
-}
