+++ presentation/src/main/java/way/presentation/feed/vo/request/FeedRequestVo.java
@@ -1,0 +1,5 @@
+package way.presentation.feed.vo.request;
+
+public class FeedRequestVo {
+
+}
