+++ presentation/src/main/java/way/presentation/feed/vo/response/FeedResponseVo.java
@@ -1,4 +1,4 @@
-package way.presentation.feed.vo.res;
+package way.presentation.feed.vo.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
